\name{outputConstructors}
\alias{outputConstructors}
\title{Write constructor code to connection}
\description{
  This writes the actual S code for the constructors
  to the specified connection and also 
}
\usage{
outputConstructors(defs, constructorFile, prefix)
}
\arguments{
  \item{defs}{}
  \item{constructorFile}{}
  \item{prefix}{}
}
\details{
}
\value{

}

\references{\url{http://www.omegahat.net/RGtk},
  \url{http://www.omegahat.net/RGtkBindingGenerator},
  \url{http://www.gtk.org}
  \url{http://www.pygtk.org}(?)
}
\author{Duncan Temple Lang <duncan@research.bell-labs.com>}

\seealso{
 \code{\link{genCode}}
}

\examples{

}
\keyword{programming} 

