\name{PrimitiveTypeCoercion}
\alias{PrimitiveTypeCoercion}
\non_function{}
\title{Table of Gtk types and S converters}
\description{
  This is a table associating Gtk types in the .defs files.
  It can be extended or modified to alter how the checking and coercion
  of arguments are converted to primitive values  before being passed
  to C routines in the Gtk interface.
}
\usage{data(PrimitiveTypeCoercion)}
\format{
  Named character vector with the elements being the names
  of S functions and the names of the elements being
  the Gtk types.
}
\examples{
}
\keyword{datasets}
