\name{RGtkBindingGenerator-internal}
\title{Internal RGtkBindingGenerator functions}
\alias{getRawDefs}
\alias{GtkPrimitiveTypes}
\aliase{PrimitiveTypeCoercion}
\usage{}
\description{
  These are internal functions for the RGtkBindingGenerator package
  and are not documented for general use.
}
\keyword{internal}